% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{call_meteor}
\alias{call_meteor}
\title{Call Meteor API Function}
\usage{
call_meteor(
  method = c("get", "post"),
  ressource = c("external", "lookup", "query", "view"),
  option = c("count", "entry", "random", "recent", "ownership", "similar", "uid", "cran",
    "doi", "instagram", "telegram", "twitter", "vk", "website"),
  type = c("Entry", "PoliticalParty", "Organization", "JournalisticBrand", "NewsSource",
    "Government", "Parliament", "Person", "Channel", "Country", "Multinational",
    "Subnational", "Archive", "Dataset", "Tool", "ScientificPublication", "Author",
    "Language", "ProgrammingLanguage", "Operation", "FileFormat", "MetaVariable",
    "ConceptVariable", "TextType", "UnitOfAnalysis", "Modality", "Collection",
    "LearningMaterial"),
  type_connector = NULL,
  audience_size = NULL,
  audience_size_connector = NULL,
  audience_size_unit = NULL,
  audience_size_unit_connector = NULL,
  audience_size_recent = NULL,
  audience_size_recent_connector = NULL,
  audience_size_recent_unit = NULL,
  audience_size_recent_unit_connector = NULL,
  channel = NULL,
  channel_connector = NULL,
  contains_ads = c("yes", "no", "non subscribers", "NA"),
  country = NULL,
  country_connector = NULL,
  countries = NULL,
  countries_connector = NULL,
  date_founded = NULL,
  date_founded_connector = NULL,
  geographic_scope = c("multinational", "national", "subnational"),
  geographic_scope_connector = NULL,
  languages = NULL,
  languages_connector = NULL,
  name_abbrev = NULL,
  name_abbrev_connector = NULL,
  ownership_kind = c("private ownership", "public ownership", "unkown", "NA"),
  payment_model = c("free", "partly free", "not free", "NA"),
  party_affiliated = c("yes", "no", "NA"),
  publication_kind = c("newspaper", "news site", "news agency", "tv show", "radio show",
    "podcast", "news blog", "alternative media", "organizational communication"),
  publication_kind_connector = NULL,
  publication_cycle = c("continuous", "daily", "multiple times per week", "weekly",
    "twice a month", "monthly", "less than monthly", "NA"),
  sources_included = NULL,
  sources_included_connector = NULL,
  subnational_scope = NULL,
  subnational_scope_connector = NULL,
  transcript_kind = NULL,
  topical_focus = c("economy", "education", "environment", "health", "media", "politics",
    "religion", "society", "science", "youth"),
  website_allows_comments = NULL,
  terms = NULL,
  is_politician = NULL,
  verified_account = NULL,
  special_interest = NULL,
  defunct = NULL,
  fulltext_available = NULL,
  graphical_user_interface = NULL,
  language_independent = NULL,
  conditions_of_access = c("NA", "free", "registration", "request", "purchase"),
  text_types = NULL,
  text_types_connector = NULL,
  text_units = NULL,
  text_units_connector = NULL,
  modalities = NULL,
  modalities_connector = NULL,
  meta_variables = NULL,
  meta_variables_connector = NULL,
  concept_variables = NULL,
  concept_variables_connector = NULL,
  date_published = NULL,
  date_published_connector = NULL,
  platforms = c("windows", "linux", "macos"),
  platforms_connector = NULL,
  programming_languages = NULL,
  programming_languages_connector = NULL,
  open_source = c("NA", "yes", "no"),
  used_for = NULL,
  used_for_connector = NULL,
  designed_for = NULL,
  designed_for_connector = NULL,
  input_file_format = NULL,
  input_file_format_connector = NULL,
  output_file_format = NULL,
  output_file_format_connector = NULL,
  author_validated = c("NA", "yes", "no"),
  validation_dataset = NULL,
  validation_dataset_connector = NULL,
  methodologies = NULL,
  methodologies_connector = NULL,
  datasets_used = NULL,
  datasets_used_connector = NULL,
  channels = NULL,
  channels_connector = NULL,
  entries_included = NULL,
  entries_included_connector = NULL,
  authors = NULL,
  authors_connector = NULL,
  add_params = NULL,
  uid = NULL,
  page = 1,
  package = NULL,
  doi = NULL,
  fresh = TRUE,
  handle = NULL,
  website = NULL,
  verbose = TRUE,
  n_max = Inf,
  n = 50,
  format = c("dataframe", "raw"),
  unnest_cutoff = NULL
)
}
\arguments{
\item{method}{A character string specifying the HTTP method to be used ("get" by default).}

\item{ressource}{A character string specifying the type of resource to interact with ("external", "lookup", "query", "view").}

\item{option}{A character string specifying additional options for the resource.}

\item{type}{A character string specifying the type of resource.}

\item{type_connector}{A character string specifying the connector for type.}

\item{audience_size}{An integer specifying the audience size.}

\item{audience_size_connector}{A character string specifying the connector for audience size.}

\item{audience_size_unit}{A character string specifying the unit for audience size.}

\item{audience_size_unit_connector}{A character string specifying the connector for audience size unit.}

\item{audience_size_recent}{An integer specifying the recent audience size.}

\item{audience_size_recent_connector}{A character string specifying the connector for recent audience size.}

\item{audience_size_recent_unit}{A character string specifying the unit for recent audience size.}

\item{audience_size_recent_unit_connector}{A character string specifying the connector for recent audience size unit.}

\item{channel}{A character string specifying the channel.}

\item{channel_connector}{A character string specifying the connector for channel.}

\item{contains_ads}{A character string specifying if the resource contains ads.}

\item{country}{A character string specifying the country.}

\item{country_connector}{A character string specifying the connector for country.}

\item{countries}{A character string specifying the countries.}

\item{countries_connector}{A character string specifying the connector for countries.}

\item{date_founded}{A character string specifying the date founded.}

\item{date_founded_connector}{A character string specifying the connector for date founded.}

\item{geographic_scope}{A character string specifying the geographic scope.}

\item{geographic_scope_connector}{A character string specifying the connector for geographic scope.}

\item{languages}{A character string specifying the languages.}

\item{languages_connector}{A character string specifying the connector for languages.}

\item{name_abbrev}{A character string specifying the name abbreviation.}

\item{name_abbrev_connector}{A character string specifying the connector for name abbreviation.}

\item{ownership_kind}{A character string specifying the ownership kind.}

\item{payment_model}{A character string specifying the payment model.}

\item{party_affiliated}{A character string specifying if party affiliated.}

\item{publication_kind}{A character string specifying the publication kind.}

\item{publication_kind_connector}{A character string specifying the connector for publication kind.}

\item{publication_cycle}{A character string specifying the publication cycle.}

\item{sources_included}{A character string specifying the sources included.}

\item{sources_included_connector}{A character string specifying the connector for sources included.}

\item{subnational_scope}{A character string specifying the subnational scope.}

\item{subnational_scope_connector}{A character string specifying the connector for subnational scope.}

\item{transcript_kind}{A character string specifying the transcript kind.}

\item{topical_focus}{A character string specifying the topical focus.}

\item{website_allows_comments}{A character string specifying if the website allows comments.}

\item{terms}{A character string specifying the terms.}

\item{is_politician}{A character string specifying if the entity is a politician.}

\item{verified_account}{A character string specifying if the account is verified.}

\item{special_interest}{A character string specifying special interests.}

\item{defunct}{A character string specifying if the resource is defunct.}

\item{fulltext_available}{A character string specifying if full text is available.}

\item{graphical_user_interface}{A character string specifying if a graphical user interface is available.}

\item{language_independent}{A character string specifying if the resource is language independent.}

\item{conditions_of_access}{A character string specifying the conditions of access.}

\item{text_types}{A character string specifying the text types.}

\item{text_types_connector}{A character string specifying the connector for text types.}

\item{text_units}{A character string specifying the text units.}

\item{text_units_connector}{A character string specifying the connector for text units.}

\item{modalities}{A character string specifying the modalities.}

\item{modalities_connector}{A character string specifying the connector for modalities.}

\item{meta_variables}{A character string specifying the meta variables.}

\item{meta_variables_connector}{A character string specifying the connector for meta variables.}

\item{concept_variables}{A character string specifying the concept variables.}

\item{concept_variables_connector}{A character string specifying the connector for concept variables.}

\item{date_published}{A character string specifying the date published.}

\item{date_published_connector}{A character string specifying the connector for date published.}

\item{platforms}{A character string specifying the platforms.}

\item{platforms_connector}{A character string specifying the connector for platforms.}

\item{programming_languages}{A character string specifying the programming languages.}

\item{programming_languages_connector}{A character string specifying the connector for programming languages.}

\item{open_source}{A character string specifying if the resource is open source.}

\item{used_for}{A character string specifying what the resource is used for.}

\item{used_for_connector}{A character string specifying the connector for what the resource is used for.}

\item{designed_for}{A character string specifying what the resource is designed for.}

\item{designed_for_connector}{A character string specifying the connector for what the resource is designed for.}

\item{input_file_format}{A character string specifying the input file format.}

\item{input_file_format_connector}{A character string specifying the connector for input file format.}

\item{output_file_format}{A character string specifying the output file format.}

\item{output_file_format_connector}{A character string specifying the connector for output file format.}

\item{author_validated}{A character string specifying if the author is validated.}

\item{validation_dataset}{A character string specifying the validation dataset.}

\item{validation_dataset_connector}{A character string specifying the connector for validation dataset.}

\item{methodologies}{A character string specifying the methodologies.}

\item{methodologies_connector}{A character string specifying the connector for methodologies.}

\item{datasets_used}{A character string specifying the datasets used.}

\item{datasets_used_connector}{A character string specifying the connector for datasets used.}

\item{channels}{A character string specifying the channels.}

\item{channels_connector}{A character string specifying the connector for channels.}

\item{entries_included}{A character string specifying the entries included.}

\item{entries_included_connector}{A character string specifying the connector for entries included.}

\item{authors}{A character string specifying the authors.}

\item{authors_connector}{A character string specifying the connector for authors.}

\item{add_params}{Additional parameters to pass.}

\item{uid}{A character string specifying the user ID.}

\item{page}{An integer specifying the page number (default is 1).}

\item{package}{A character string specifying the package name.}

\item{doi}{A character string specifying the DOI.}

\item{fresh}{A logical value indicating freshness (default is TRUE).}

\item{handle}{A character string specifying the handle.}

\item{website}{A character string specifying the website URL.}

\item{verbose}{A logical value indicating whether to display progress messages (default is TRUE).}

\item{n_max}{An integer specifying the maximum number of results to fetch (default is Inf).}

\item{n}{An integer specifying the number of results per page (default is 50).}

\item{format}{A character string indicating the format of the output ("dataframe" or "raw").}

\item{unnest_cutoff}{An integer specifying the cutoff point for unnesting the results.}
}
\value{
The retrieved data in the specified format.
}
\description{
This function performs API calls to the Meteor API with customizable parameters for both GET and POST requests.
}
\details{
Currently, the package supports only API endpoints to retrieve data, i.e. not calls to edit entries.
Full documentation can be found here: \url{https://meteor.opted.eu/api/swagger}
}
\examples{
np <- call_meteor(method = "get",
                  ressource = "query",
                  type = "NewsSource",
                  countries = c("0x1b", "0x2f"),
                  format = "dataframe",
                  publication_kind = "newspaper",
                  n_max = 10,
                  n = 10)

}
