% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{callAPI}
\alias{callAPI}
\title{Call API Function}
\usage{
callAPI(url, method = c("get", "post"), body = NULL)
}
\arguments{
\item{url}{A character string representing the URL of the API call.}

\item{method}{A character string indicating the HTTP method to be used ("get" or "post").}

\item{body}{A list representing the body of the request, required for POST requests.}
}
\value{
The response from the API call.
}
\description{
This function makes a call to the specified API URL using the provided HTTP method and returns the response.
}
\examples{
res <- callAPI(method = "get",
              url = generateURL(method = "get", ressource = "view", option = "random"))

}
