% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coalesce_join}
\alias{coalesce_join}
\title{coalesce_join
ajout information a data frame existant}
\usage{
coalesce_join(dataset_ori, dataset_add, valid_cols, join_by, verbose = FALSE)
}
\arguments{
\item{dataset_ori}{; dataset a complementer}

\item{dataset_add}{: dataset contenant les informations a ajouter}

\item{valid_cols}{: vecteur contenant les champs attendus dans dataset_ori}

\item{join_by}{: vecteur contenant les champs utilises pour joindre les datasets}

\item{verbose}{: booleen servant a masquer / afficher les messages}
}
\description{
coalesce_join
ajout information a data frame existant
}
