% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_data.R
\docType{data}
\name{covid_data}
\alias{covid_data}
\title{Real world metabolomics data set to test MeTEor}
\format{
A data frame with 21593 rows and 7 variables
}
\usage{
data(covid_data)
}
\description{
An example real world metabolomics data set from the study "Plasma Metabolome Alterations Discriminate between COVID-19 and Non-COVID-19 Pneumonia" to test MeTEor.
}
\references{
More, T. H., Mozafari, B., Märtens, A., Herr, C., Lepper, P. M., Danziger, G., ... & Bals, R. (2022). Plasma Metabolome Alterations Discriminate between COVID-19 and Non-COVID-19 Pneumonia. Metabolites, 12(11), 1058.
}
\keyword{datasets}
