% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{csv_to_rds}
\alias{csv_to_rds}
\title{CSV to RDS files transformation}
\usage{
csv_to_rds(rds_directory = getwd(), file_lst = NULL, ...)
}
\arguments{
\item{rds_directory}{: file path where rds files will be stored}

\item{file_lst}{: list of csv files to convert}

\item{...}{: optional arguments}
}
\description{
CSV to RDS files transformation
}
\examples{
dir_csv <- "f:/r/netatmo/csv"
dir_rds <- "e:/work/netatmo2"
file_lst <- netatmo_file_to_load(dir_csv, pattern = list("Indoor"))
csv_to_rds(dir_rds, file_lst)
file_lst <- netatmo_file_to_load(dir_csv, pattern = list("Indoor", "Outdoor"))
csv_to_rds(dir_rds, file_lst)
}
