% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_check.R
\name{data_check}
\alias{data_check}
\title{Check Data for MeTEor}
\usage{
data_check(data)
}
\arguments{
\item{data}{A data frame containing the input data.}
}
\value{
No explicit return value. The function stops with an error if any check fails.
}
\description{
This function performs several checks on the input data to ensure it meets the expected
requirements for analysis. It checks column names, data types, and performs additional
validations.
}
\examples{
\dontrun{
# Assuming covid_data is your data frame
data_check(covid_data)
}

}
