% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{day_length}
\alias{day_length}
\title{Calculate the day length at given DOY at given latitude.}
\usage{
day_length(DOY, latitude)
}
\arguments{
\item{DOY}{day of year}

\item{latitude}{Geographical latitude in degrees}
}
\value{
Day length hours
}
\description{
Calculate the day length at given DOY at given latitude.
}
