% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Est_SampleStructure.R
\name{est_SS}
\alias{est_SS}
\title{A function estimate parameters of sample structure by LD score regression.}
\usage{
est_SS(
  dat1,
  dat2,
  trait1.name = "exposure",
  trait2.name = "outcome",
  LDSC = T,
  h2.fix.intercept = F,
  ldscore.dir = NULL,
  ld = NULL,
  M = NULL
)
}
\arguments{
\item{dat1}{formatted GWAS summary-level data for exposure.}

\item{dat2}{formatted GWAS summary-level data for outcome.}

\item{trait1.name}{specify the name of exposure, default `exposure`.}

\item{trait2.name}{specify the name of outcome, default `outcome`.}

\item{LDSC}{logical, whether to run LD score regression, default `TRUE`. If `FALSE`, the function will not give the parameter estimates.}

\item{h2.fix.intercept}{logical, whether to fix LD score regression intercept to 1, default `FALSE`.}

\item{ldscore.dir}{specify the path to the LD score files.}

\item{ld}{specify the data.frame for the ld scores of SNPs if available and setting ldscore.dir =NULL.}

\item{M}{specify the number of SNPs used for calculating LD score.}
}
\value{
List with the following elements:
\describe{
\item{Omega}{the estimated Omega matrix capturing the effects of sample structure}
\item{Omega.se}{the estimated matrix consists of the standard errors of the intercept estimates obtained from LD score regression}
}
}
\description{
Format GWAS summary data for calculating Sample structure.
This function is adapted from the est_paras() function in YangLabHKUST/MR-APSS.
}
