% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Simulated longitudinal metabolomics data}
\format{
A data frame with 180000 rows and 6 variables
}
\usage{
data(example_data)
}
\description{
Simulated longitudinal metabolomics dataset containing
a patient identifier (id), timepoint of measurement (time),
metabolites (metabolites), measurement values (values),
categorical variables (treatment, sex).
}
\keyword{datasets}
