% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extraterrestrial_radiation}
\alias{extraterrestrial_radiation}
\title{Extraterrestrial radiation at given DOY at given latitude.}
\usage{
extraterrestrial_radiation(DOY, lat)
}
\arguments{
\item{DOY}{day of year}

\item{lat}{Geographical latitude in degree}
}
\value{
Average extraterrestrial radiation in J/s/m^2, sometimes called
srad.
}
\description{
Extraterrestrial radiation at given DOY at given latitude.
}
