% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_get_site.R
\name{forecast_get_site}
\alias{forecast_get_site}
\title{Retrieve the 5 day weather forecast for a site}
\usage{
forecast_get_site(site_id)
}
\arguments{
\item{site_id}{a valid met office site id}
}
\description{
Requests three-hourly forecast from the met office API in xml format and returns a \code{tibble}
}
\examples{
forecast_get_site(3840)
}
