% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_list_sites.R
\name{forecast_list_sites}
\alias{forecast_list_sites}
\title{List available locations for forecasts}
\usage{
forecast_list_sites()
}
\description{
Requests locations from the met office API in xml format and returns a \code{tibble}
}
\examples{
forecast_list_sites()
}
