% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format_data.R
\name{format_data}
\alias{format_data}
\title{Format GWAS summary data for calculating Sample structure.}
\usage{
format_data(
  dat,
  snps.merge = NULL,
  snps.remove = NULL,
  snp_col = "SNP",
  b_col = "b",
  se_col = "se",
  freq_col = "freq",
  A1_col = "A1",
  A2_col = "A2",
  p_col = "p",
  n_col = "n",
  info_col = "INFO",
  chi2_max = NULL,
  min_freq = 0.05
)
}
\arguments{
\item{dat}{data.frame with at least SNP A1 A2 signed statistics for calculating z scores and sample size.}

\item{snps.merge}{data.frame with SNPs to extract. The data frame must have headers: SNP A1 and A2. For example, the hapmap3 SNP list.}

\item{snps.remove}{a set of SNPs that needed to be removed. For example, SNPs in MHC region.}

\item{snp_col}{name of a column with rs numbers. The default is \code{SNP}.}

\item{b_col}{name of a column with effect sizes. The default is \code{b}.}

\item{se_col}{name of a column with standard errors. The default is \code{se}.}

\item{freq_col}{name of a column with effect allele frequencies. The default is \code{freq}.}

\item{A1_col}{name of a column with effect alleles. The default is \code{A1}.}

\item{A2_col}{name of a column with the other alleles. The default is \code{A2}.}

\item{p_col}{name of a column with p-values. The default is \code{p}.}

\item{n_col}{name of a column with sample sizes. The default is \code{n}.}

\item{info_col}{name of a column with imputation INFO. The default is \code{Info}.}

\item{chi2_max}{SNPs with tested chi^2 statistics large than chi2_max will be removed. The default is \code{80}}

\item{min_freq}{SNPs with allele frequency less than min_freq will be removed. The default is \code{0.05}.}

\item{n}{sample size. If the column for sample size is not available, users can use argument "n" to specify the total sample size for each SNP.}
}
\value{
a data frame with headers: SNP: rs number; A1: effect allele; A2: the other allele;
Z: Z score;  N: sample size;  chi2: chi-square statistics;  P: p-value.
}
\description{
Format GWAS summary data for calculating Sample structure.
This function is adapted from the format_data() function in MRCIEU/TwoSampleMR.
}
