\name{FSE weather}

\alias{readFSEwth}
\alias{writeFSEwth}
\alias{default_weather} 


\title{
Read FSE formatted weather data
}

\description{
Read or write FSE formatted weather data
}

\usage{
readFSEwth(f)
default_weather()
}

\arguments{
\item{f}{character. filename}
}

\value{
data.frame
}
