% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{generate_body}
\alias{generate_body}
\title{Generate Body for API Call}
\usage{
generate_body(
  package = NULL,
  doi = NULL,
  fresh = TRUE,
  handle = NULL,
  body_url = NULL
)
}
\arguments{
\item{package}{A character string specifying the package name.}

\item{doi}{A character string specifying the DOI identifier.}

\item{fresh}{A logical value indicating whether the data should be fetched fresh.}

\item{handle}{A character string specifying the social media handle.}

\item{body_url}{A character string specifying the URL.}
}
\value{
A list representing the generated body for the API call.
}
\description{
This function generates the body for a POST method API call based on the specified parameters.
}
\examples{
body <- generate_body(package = "dplyr")

}
