% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{generateURL}
\alias{generateURL}
\title{Generate URL for OPTED METEOR API}
\usage{
generateURL(
  method = c("get", "post"),
  ressource = c("external", "lookup", "query", "view"),
  option = c("count", "entry", "random", "recent", "ownership", "similar", "uid", "cran",
    "doi", "instagram", "telegram", "twitter", "vk", "website"),
  n = 50,
  type = c("Entry", "PoliticalParty", "Organization", "JournalisticBrand", "NewsSource",
    "Government", "Parliament", "Person", "Channel", "Country", "Multinational",
    "Subnational", "Archive", "Dataset", "Tool", "ScientificPublication", "Author",
    "Language", "ProgrammingLanguage", "Operation", "FileFormat", "MetaVariable",
    "ConceptVariable", "TextType", "UnitOfAnalysis", "Modality", "Collection",
    "LearningMaterial"),
  type_connector = NULL,
  audience_size = NULL,
  audience_size_connector = NULL,
  audience_size_unit = NULL,
  audience_size_unit_connector = NULL,
  audience_size_recent = NULL,
  audience_size_recent_connector = NULL,
  audience_size_recent_unit = NULL,
  audience_size_recent_unit_connector = NULL,
  channel = NULL,
  channel_connector = NULL,
  contains_ads = c("yes", "no", "non subscribers", "NA"),
  country = NULL,
  country_connector = NULL,
  countries = NULL,
  countries_connector = NULL,
  date_founded = NULL,
  date_founded_connector = NULL,
  geographic_scope = c("multinational", "national", "subnational"),
  geographic_scope_connector = NULL,
  languages = NULL,
  languages_connector = NULL,
  name_abbrev = NULL,
  name_abbrev_connector = NULL,
  ownership_kind = c("private ownership", "public ownership", "unkown", "NA"),
  payment_model = c("free", "partly free", "not free", "NA"),
  party_affiliated = c("yes", "no", "NA"),
  publication_kind = c("newspaper", "news site", "news agency", "tv show", "radio show",
    "podcast", "news blog", "alternative media", "organizational communication"),
  publication_kind_connector = NULL,
  publication_cycle = c("continuous", "daily", "multiple times per week", "weekly",
    "twice a month", "monthly", "less than monthly", "NA"),
  sources_included = NULL,
  sources_included_connector = NULL,
  subnational_scope = NULL,
  subnational_scope_connector = NULL,
  transcript_kind = NULL,
  topical_focus = c("economy", "education", "environment", "health", "media", "politics",
    "religion", "society", "science", "youth"),
  website_allows_comments = NULL,
  terms = NULL,
  is_politician = NULL,
  verified_account = NULL,
  special_interest = NULL,
  defunct = NULL,
  fulltext_available = NULL,
  graphical_user_interface = NULL,
  language_independent = NULL,
  conditions_of_access = c("NA", "free", "registration", "request", "purchase"),
  text_types = NULL,
  text_types_connector = NULL,
  text_units = NULL,
  text_units_connector = NULL,
  modalities = NULL,
  modalities_connector = NULL,
  meta_variables = NULL,
  meta_variables_connector = NULL,
  concept_variables = NULL,
  concept_variables_connector = NULL,
  date_published = NULL,
  date_published_connector = NULL,
  platforms = c("windows", "linux", "macos"),
  platforms_connector = NULL,
  programming_languages = NULL,
  programming_languages_connector = NULL,
  open_source = c("NA", "yes", "no"),
  used_for = NULL,
  used_for_connector = NULL,
  designed_for = NULL,
  designed_for_connector = NULL,
  input_file_format = NULL,
  input_file_format_connector = NULL,
  output_file_format = NULL,
  output_file_format_connector = NULL,
  author_validated = c("NA", "yes", "no"),
  validation_dataset = NULL,
  validation_dataset_connector = NULL,
  methodologies = NULL,
  methodologies_connector = NULL,
  datasets_used = NULL,
  datasets_used_connector = NULL,
  channels = NULL,
  channels_connector = NULL,
  entries_included = NULL,
  entries_included_connector = NULL,
  authors = NULL,
  authors_connector = NULL,
  add_params = NULL,
  uid = NULL,
  page = 1
)
}
\arguments{
\item{method}{The HTTP method to use. Default is "get".}

\item{ressource}{The resource to access.}

\item{option}{The option to specify within the resource.}

\item{n}{Number of results. Default is 50.}

\item{type}{Array of types.}

\item{type_connector}{Connector for type array.}

\item{audience_size}{Audience size.}

\item{audience_size_connector}{Connector for audience size.}

\item{audience_size_unit}{Unit for audience size.}

\item{audience_size_unit_connector}{Connector for audience size unit.}

\item{audience_size_recent}{Recent audience size.}

\item{audience_size_recent_connector}{Connector for recent audience size.}

\item{audience_size_recent_unit}{Unit for recent audience size.}

\item{audience_size_recent_unit_connector}{Connector for recent audience size unit.}

\item{channel}{Channel.}

\item{channel_connector}{Connector for channel.}

\item{contains_ads}{Indicator for ads.}

\item{country}{Country.}

\item{country_connector}{Connector for country.}

\item{countries}{Array of countries.}

\item{countries_connector}{Connector for countries.}

\item{date_founded}{Date founded.}

\item{date_founded_connector}{Connector for date founded.}

\item{geographic_scope}{Geographic scope.}

\item{geographic_scope_connector}{Connector for geographic scope.}

\item{languages}{Array of languages.}

\item{languages_connector}{Connector for languages.}

\item{name_abbrev}{Name abbreviation.}

\item{name_abbrev_connector}{Connector for name abbreviation.}

\item{ownership_kind}{Ownership kind.}

\item{payment_model}{Payment model.}

\item{party_affiliated}{Party affiliation.}

\item{publication_kind}{Publication kind.}

\item{publication_kind_connector}{Connector for publication kind.}

\item{publication_cycle}{Publication cycle.}

\item{sources_included}{Included sources.}

\item{sources_included_connector}{Connector for included sources.}

\item{subnational_scope}{Subnational scope.}

\item{subnational_scope_connector}{Connector for subnational scope.}

\item{transcript_kind}{Transcript kind.}

\item{topical_focus}{Topical focus.}

\item{website_allows_comments}{Indicator for allowing comments on website.}

\item{terms}{Search terms.}

\item{is_politician}{Indicator for politician.}

\item{verified_account}{Indicator for verified account.}

\item{special_interest}{Special interest.}

\item{defunct}{Indicator for defunct.}

\item{fulltext_available}{Indicator for fulltext availability.}

\item{graphical_user_interface}{Indicator for graphical user interface.}

\item{language_independent}{Indicator for language independence.}

\item{conditions_of_access}{Conditions of access.}

\item{text_types}{Array of text types.}

\item{text_types_connector}{Connector for text types.}

\item{text_units}{Array of text units.}

\item{text_units_connector}{Connector for text units.}

\item{modalities}{Array of modalities.}

\item{modalities_connector}{Connector for modalities.}

\item{meta_variables}{Array of meta variables.}

\item{meta_variables_connector}{Connector for meta variables.}

\item{concept_variables}{Array of concept variables.}

\item{concept_variables_connector}{Connector for concept variables.}

\item{date_published}{Date published.}

\item{date_published_connector}{Connector for date published.}

\item{platforms}{Array of platforms.}

\item{platforms_connector}{Connector for platforms.}

\item{programming_languages}{Array of programming languages.}

\item{programming_languages_connector}{Connector for programming languages.}

\item{open_source}{Indicator for open source.}

\item{used_for}{Used for.}

\item{used_for_connector}{Connector for used for.}

\item{designed_for}{Designed for.}

\item{designed_for_connector}{Connector for designed for.}

\item{input_file_format}{Input file format.}

\item{input_file_format_connector}{Connector for input file format.}

\item{output_file_format}{Output file format.}

\item{output_file_format_connector}{Connector for output file format.}

\item{author_validated}{Indicator for author validation.}

\item{validation_dataset}{Validation dataset.}

\item{validation_dataset_connector}{Connector for validation dataset.}

\item{methodologies}{Array of methodologies.}

\item{methodologies_connector}{Connector for methodologies.}

\item{datasets_used}{Array of datasets used.}

\item{datasets_used_connector}{Connector for datasets used.}

\item{channels}{Array of channels.}

\item{channels_connector}{Connector for channels.}

\item{entries_included}{Included entries.}

\item{entries_included_connector}{Connector for included entries.}

\item{authors}{Array of authors.}

\item{authors_connector}{Connector for authors.}

\item{add_params}{Additional parameters to append to the URL.}

\item{uid}{User ID.}

\item{page}{Page number.}
}
\value{
A character string representing the generated URL.
}
\description{
This function generates a URL for accessing the OPTED METEOR API based on specified parameters.
}
\details{
Currently, the package supports only API endpoints to retrieve data, i.e. not calls to edit entries.
Full documentation can be found here: \url{https://meteor.opted.eu/api/swagger}
}
\examples{
generateURL(method = "get", ressource = "view", option = "random")

}
