% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_generic}
\alias{get_generic}
\title{Generic GET Function}
\usage{
get_generic(
  method = "get",
  ressource = c("external", "lookup", "query", "view"),
  verbose = TRUE,
  n_max = Inf,
  n = 50,
  format = c("dataframe", "raw"),
  ...
)
}
\arguments{
\item{method}{A character string specifying the HTTP method to be used ("get" by default).}

\item{ressource}{A character string specifying the type of resource to retrieve data from ("external", "lookup", "query", "view").}

\item{verbose}{Logical indicating whether to display progress messages (default is TRUE).}

\item{n_max}{An integer specifying the maximum number of results to fetch (default is Inf).}

\item{n}{An integer specifying the number of results per page (default is 50).}

\item{format}{A character string indicating the format of the output ("dataframe" or "raw").}

\item{...}{Additional arguments passed to the function for generating the URL.}
}
\value{
The retrieved data in the specified format.
}
\description{
This function performs a GET request to retrieve data from the specified API resource using pagination.
}
\examples{
news_sources <- get_generic(method = "get",
                           ressource = "query",
                           type = "NewsSource",
                           countries = c("0x1b"),
                           geographic_scope = "national",
                           n_max = 10,
                           n = 10,
                           format = "dataframe"
                           )

}
