% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico_temperatura_mensual.R
\name{grafico_temperatura_mensual}
\alias{grafico_temperatura_mensual}
\title{Title Grafico de Temperatura Promedio Mensual}
\usage{
grafico_temperatura_mensual(
  datos,
  colores = c("blue", "red"),
  titulo = "Temperatura Promedio Mensual"
)
}
\arguments{
\item{datos}{dataframe con las columnas fecha, id y temperatura_abrigo_150cm}

\item{colores}{colores de las lineas del grafico}

\item{titulo}{titulo del grafico}
}
\value{
grafico de lineas con la temperatura promedio mensual por id
}
\description{
Title Grafico de Temperatura Promedio Mensual
}
\examples{
  grafico_temperatura_mensual(NH0046)
}
