% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{hargreaves_et0_extraterrestrial}
\alias{hargreaves_et0_extraterrestrial}
\title{FAO Hargreaves equation, alternative implementation}
\usage{
hargreaves_et0_extraterrestrial(T_mean, T_max, T_min, r_extraterrestrial)
}
\arguments{
\item{T_mean}{Daily average temperature in °C}

\item{T_max}{Daily maximum temperature in °C}

\item{T_min}{Daily minimum temperature in °C}

\item{r_extraterrestrial}{Daily average extraterrestrial radiation in W/m^2}
}
\description{
Estimate reference evapotranspiration based on temperature and radiation
data. See equation 52 in Chapter 3 of "Crop evapotranspiration -
Guidelines for computing crop water requirements - FAO Irrigation and
drainage paper 56"
}
