% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{hargreaves_et0_terrestrial}
\alias{hargreaves_et0_terrestrial}
\title{Original Hargreaves-Samani equation}
\usage{
hargreaves_et0_terrestrial(T_mean, r_terrestrial)
}
\arguments{
\item{T_mean}{Daily average temperature in °C}

\item{r_terrestrial}{Daily average terrestrial radiation in W/m^2}
}
\value{
Reference evapotranspiration in mm/day
}
\description{
Original Hargreaves-Samani equation
}
