% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{import_ext_rain}
\alias{import_ext_rain}
\title{Import rainfall from external module}
\usage{
import_ext_rain(csv_name = NULL, ...)
}
\arguments{
\item{csv_name}{: file to import}

\item{...}{: optional arguments}
}
\value{
: a R tibble containing external rainfall
}
\description{
Import rainfall from external module
}
\examples{
csv_name <- "f:/r/netatmo/csv/Pluviometre_04_2017.csv"
tmp <- import_ext_rain(csv_name = csv_name)
}
