% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{import_ext_temp_hum}
\alias{import_ext_temp_hum}
\title{Import temperature and humidity from external module}
\usage{
import_ext_temp_hum(csv_name = NULL, ...)
}
\arguments{
\item{csv_name}{: file to import}

\item{...}{: optional arguments}
}
\value{
: a R tibble containing external temperature and humidity
}
\description{
Import temperature and humidity from external module
}
\examples{
csv_name <- "f:/r/netatmo/csv/Exterieur_04_2017.csv"
boo <- import_ext_temp_hum(csv_name = csv_name)
}
