% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{import_ext_wind}
\alias{import_ext_wind}
\title{Import wind from external module}
\usage{
import_ext_wind(csv_name = NULL, ...)
}
\arguments{
\item{csv_name}{: file to import}

\item{...}{: optional arguments}
}
\value{
: a R tibble containing external wind and gust
}
\description{
Import wind from external module
}
\examples{
csv_name <- "f:/r/netatmo/csv/Anemometre_04_2017.csv"
boo <- import_ext_wind(csv_name = csv_name)
}
