% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{import_int}
\alias{import_int}
\title{Import temperature, humidity, c02, noise and pressure from internal module}
\usage{
import_int(csv_name = NULL, ...)
}
\arguments{
\item{csv_name}{: file to import}

\item{...}{: optional arguments}
}
\value{
: a R tibble containing internal temperature, humidity, c02, noise and pressure
}
\description{
Import temperature, humidity, c02, noise and pressure from internal module
}
\examples{
csv_name <- "f:/r/netatmo/csv/Indoor_04_2017.csv"
tmp <- import_ext_wind(csv_name = csv_name)
}
