% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{inverse_relative_distance}
\alias{inverse_relative_distance}
\title{Inverse relative distance between Earth and Sun}
\usage{
inverse_relative_distance(DOY)
}
\arguments{
\item{DOY}{day of year}
}
\description{
Inverse relative distance between Earth and Sun
}
