% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_valid_dataset}
\alias{is_valid_dataset}
\title{is_valid_df
controle validite data frame}
\usage{
is_valid_dataset(dataset, valid_cols)
}
\arguments{
\item{dataset}{: dataset a tester}

\item{valid_cols}{: vecteur contenant les champs attendus dans dataset}
}
\description{
is_valid_df
controle validite data frame
}
