% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leer_datos.R
\name{leer_estacion}
\alias{leer_estacion}
\title{Leer datos de una estacion meteorologica}
\usage{
leer_estacion(id, ruta)
}
\arguments{
\item{id}{El id de la estacion a leer}

\item{ruta}{Ruta donde se descargan o leen los datos}
}
\value{
Un dataframe con los datos de la estacion en la ruta indicada
}
\description{
Leer datos de una estacion meteorologica
}
\examples{
  ruta_tmp <- file.path(tempdir(), "NH0472.csv")
  df <- leer_estacion("NH0472", ruta_tmp)
  head(df)

}
