% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_csv}
\alias{load_csv}
\title{load CSV data}
\usage{
load_csv(file_lst = NULL, ...)
}
\arguments{
\item{file_lst}{: list of csv files to convert}

\item{...}{: optional arguments}
}
\description{
load CSV data
}
\examples{
dir_csv <- "f:/r/netatmo/csv"
file_lst <- netatmo_file_to_load(dir_csv, pattern = list("indoor"))
db_lst <- load_csv(file_lst)
listviewer::jsonedit(db_lst)
db_lst$db_netatmo_indoor
file_lst <- netatmo_file_to_load(dir_csv, pattern = list("indoor", "outdoor"))
db_lst <- load_csv(file_lst)
listviewer::jsonedit(db_lst)
db_lst$db_netatmo_indoor
db_lst$db_netatmo_outdoor
}
