% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_rds}
\alias{load_rds}
\title{load RDS data}
\usage{
load_rds(rds_directory = getwd(), ...)
}
\arguments{
\item{rds_directory}{: file path where rds files to load are be stored}

\item{...}{: optional arguments}
}
\description{
load RDS data
}
\examples{
dir_csv <- "f:/r/netatmo/csv"
dir_rds <- "e:/work/netatmo2"
file_lst <- netatmo_file_to_load(dir_csv, pattern = list("indoor"))
csv_to_rds(dir_rds, file_lst)
db_lst <- load_rds(dir_rds)
listviewer::jsonedit(db_lst)
db_lst$db_netatmo_indoor
file_lst <- netatmo_file_to_load(dir_csv, pattern = list("indoor", "outdoor"))
csv_to_rds(dir_rds, file_lst)
db_lst <- load_rds(dir_rds)
listviewer::jsonedit(db_lst)
db_lst$db_netatmo_indoor
db_lst$db_netatmo_outdoor
}
