% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteoR-package.R
\docType{package}
\name{meteoR-package}
\alias{meteoR}
\alias{meteoR-package}
\title{meteoR: Meteorological calculations.}
\description{
Contains functions frequently used when meteorological or climatic variables have to be (i) approximately derived from other variables, (ii) corrected or gap-filled.
}
\author{
\strong{Maintainer}: Kevin Kramer \email{kevin.pasqual.kramer@protonmail.ch} (\href{https://orcid.org/0000-0001-5523-6924}{ORCID}) [copyright holder]

}
\keyword{internal}
