% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/METEOR.R
\name{METEOR}
\alias{METEOR}
\title{The function of METEOR method}
\usage{
METEOR(
  Zscore,
  Sigma,
  N,
  Omega_est,
  Omega_se,
  Gibbsnumber = 1000,
  burninproportion = 0.2,
  lambda = 0.1,
  pi_beta_shape = 0.5,
  pi_beta_scale = 4.5,
  pi_1_shape = 0.5,
  pi_1_scale = 1.5,
  pi_0_shape = 0.05,
  pi_0_scale = 9.95
)
}
\arguments{
\item{Zscore}{the Zscore with k+1 rows and p columns of the SNP effect size matrix for the exposure and k outcomes}

\item{Sigma}{the LD matrix for the SNPs selected from the exposure can be obtained by using the weighted average LD matrix from k outcomes. If individual data is unavailable, the LD matrix can also be derived from a reference panel.}

\item{N}{the sample size of exposure and k outcomes GWASs}

\item{Omega_est}{the correlation matrix derived by LDSC}

\item{Omega_se}{the standard errors of elements of Omega derived by LDSC}

\item{Gibbsnumber}{the number of Gibbs sampling iterations with the default to be 1000}

\item{burninproportion}{the proportion to burn in from Gibbs sampling iterations, with default to be 0.2}

\item{lambda}{the tuning parameter used to ensure that the correlation matrix is invertible}

\item{pi_beta_shape}{the prior shape paramter for pi_beta with the default to be 0.5}

\item{pi_beta_scale}{the prior scale paramter for pi_beta with the default to be 4.5}

\item{pi_1_shape}{the prior shape paramter for pi_1 with the default to be 0.5}

\item{pi_1_scale}{the prior scale paramter for pi_1 with the default to be 1.5}

\item{pi_0_shape}{the prior shape paramter for pi_0 with the default to be 0.05}

\item{pi_0_scale}{the prior scale paramter for pi_0 with the default to be 9.95}
}
\value{
A list of estimated parameters for the causal effect test 
\item{causal_effect}{The estimate of k causal effects}
\item{causal_pvalue_single}{The p values for the causal effects in single tests}
\item{causal_pvalue_overall}{The p value for the causal effects in overall test}
\item{causal_sd}{The standard deviation for the causal effects}
\item{causal_cov}{The covariance for the causal effects}
\item{sigmabeta}{The variance estimate for the SNP effect sizes on the exposure}
\item{sigmaeta}{The variance estimates for the horizontal pleiotropy effects}
\item{Omega}{The correlation matrix for one exposure and k outcomes}
\item{pi_beta}{The proportion of selected SNPs, which show non-zero effects on exposure}
\item{pi_1}{The proportion of selected SNPs showing horizontal pleiotropy}
\item{pi_0}{The proportion of non-selected SNPs showing horizontal pleiotropy}
}
\description{
METEOR is able to model multiple outcomes simultaneously with an initial set of candidate SNP instruments that are in high LD with each other and perform automated instrument selection to identify suitable SNPs to serve as instrumental variables.
}
