% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{METEOR_CPP}
\alias{METEOR_CPP}
\title{METEOR}
\usage{
METEOR_CPP(
  Zscore_in,
  Sigma_in,
  N_in,
  p_in,
  k_in,
  Omega_in,
  Gibbsnumberin,
  burninproportion,
  pi_beta_shape_in,
  pi_beta_scale_in,
  pi_1_shape_in,
  pi_1_scale_in,
  pi_0_shape_in,
  pi_0_scale_in
)
}
\description{
METEOR
}
