% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{netatmo_file_to_load}
\alias{netatmo_file_to_load}
\title{List files to convert}
\usage{
netatmo_file_to_load(directory = NULL, pattern = NULL, ...)
}
\arguments{
\item{directory}{: file path where csv files are stored}

\item{pattern}{: list of module patterns to load}

\item{...}{: optional arguments}
}
\value{
: one or two lists of csv files to load
}
\description{
List files to convert
}
\examples{
directory <- "f:/r/netatmo/csv"
file_lst <- netatmo_file_to_load(directory, pattern = list("indoor", "outdoor"))
listviewer::jsonedit(file_lst)
file_lst <- netatmo_file_to_load(directory, pattern = list("indoor"))
listviewer::jsonedit(file_lst)
file_lst <- netatmo_file_to_load(directory, pattern = list("outdoor"))
listviewer::jsonedit(file_lst)
}
