% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_sensor_int.R
\name{netatmo_indoor}
\alias{netatmo_indoor}
\title{netatmo_indoor object R6}
\description{
netatmo_indoor object R6

netatmo_indoor object R6
}
\examples{
dir_csv <- "f:/r/netatmo/csv"
file_lst <- netatmo_file_to_load(dir_csv, pattern = list("indoor", "outdoor"))
db_lst <- load_csv(file_lst)
tmp <- netatmo_indoor_obj$new(db_lst$db_netatmo_indoor)
}
\section{Super class}{
\code{\link[meteoR:default_data_frame]{meteoR::default_data_frame}} -> \code{netatmo_indoor}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-netatmo_indoor-new}{\code{netatmo_indoor$new()}}
\item \href{#method-netatmo_indoor-finalize}{\code{netatmo_indoor$finalize()}}
\item \href{#method-netatmo_indoor-summary}{\code{netatmo_indoor$summary()}}
\item \href{#method-netatmo_indoor-clone}{\code{netatmo_indoor$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="meteoR" data-topic="default_data_frame" data-id="head"><a href='../../meteoR/html/default_data_frame.html#method-default_data_frame-head'><code>meteoR::default_data_frame$head()</code></a></li>
<li><span class="pkg-link" data-pkg="meteoR" data-topic="default_data_frame" data-id="print"><a href='../../meteoR/html/default_data_frame.html#method-default_data_frame-print'><code>meteoR::default_data_frame$print()</code></a></li>
<li><span class="pkg-link" data-pkg="meteoR" data-topic="default_data_frame" data-id="subset"><a href='../../meteoR/html/default_data_frame.html#method-default_data_frame-subset'><code>meteoR::default_data_frame$subset()</code></a></li>
<li><span class="pkg-link" data-pkg="meteoR" data-topic="default_data_frame" data-id="tail"><a href='../../meteoR/html/default_data_frame.html#method-default_data_frame-tail'><code>meteoR::default_data_frame$tail()</code></a></li>
<li><span class="pkg-link" data-pkg="meteoR" data-topic="default_data_frame" data-id="view"><a href='../../meteoR/html/default_data_frame.html#method-default_data_frame-view'><code>meteoR::default_data_frame$view()</code></a></li>
<li><span class="pkg-link" data-pkg="meteoR" data-topic="default_data_frame" data-id="write_file"><a href='../../meteoR/html/default_data_frame.html#method-default_data_frame-write_file'><code>meteoR::default_data_frame$write_file()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-netatmo_indoor-new"></a>}}
\if{latex}{\out{\hypertarget{method-netatmo_indoor-new}{}}}
\subsection{Method \code{new()}}{
Create a new indoor object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{netatmo_indoor$new(dataset)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset}}{indoor dataset file}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `netatmo_indoor` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-netatmo_indoor-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-netatmo_indoor-finalize}{}}}
\subsection{Method \code{finalize()}}{
finalize function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{netatmo_indoor$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-netatmo_indoor-summary"></a>}}
\if{latex}{\out{\hypertarget{method-netatmo_indoor-summary}{}}}
\subsection{Method \code{summary()}}{
Summary of netamo_indoor object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{netatmo_indoor$summary(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{options of head function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-netatmo_indoor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-netatmo_indoor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{netatmo_indoor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
