% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_sensor_out.R
\name{netatmo_outdoor}
\alias{netatmo_outdoor}
\title{netatmo_outdoor object R6}
\description{
netatmo_outdoor object R6

netatmo_outdoor object R6
}
\examples{
dir_csv <- "f:/r/netatmo/csv"
file_lst <- netatmo_file_to_load(dir_csv, pattern = list("indoor", "outdoor"))
db_lst <- load_csv(file_lst)
tmp <- netatmo_outdoor$new(db_lst$db_netatmo_outdoor)
}
\section{Super class}{
\code{\link[meteoR:default_data_frame]{meteoR::default_data_frame}} -> \code{netatmo_outdoor}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-netatmo_outdoor-summary}{\code{netatmo_outdoor$summary()}}
\item \href{#method-netatmo_outdoor-new}{\code{netatmo_outdoor$new()}}
\item \href{#method-netatmo_outdoor-finalize}{\code{netatmo_outdoor$finalize()}}
\item \href{#method-netatmo_outdoor-clone}{\code{netatmo_outdoor$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="meteoR" data-topic="default_data_frame" data-id="head"><a href='../../meteoR/html/default_data_frame.html#method-default_data_frame-head'><code>meteoR::default_data_frame$head()</code></a></li>
<li><span class="pkg-link" data-pkg="meteoR" data-topic="default_data_frame" data-id="print"><a href='../../meteoR/html/default_data_frame.html#method-default_data_frame-print'><code>meteoR::default_data_frame$print()</code></a></li>
<li><span class="pkg-link" data-pkg="meteoR" data-topic="default_data_frame" data-id="subset"><a href='../../meteoR/html/default_data_frame.html#method-default_data_frame-subset'><code>meteoR::default_data_frame$subset()</code></a></li>
<li><span class="pkg-link" data-pkg="meteoR" data-topic="default_data_frame" data-id="tail"><a href='../../meteoR/html/default_data_frame.html#method-default_data_frame-tail'><code>meteoR::default_data_frame$tail()</code></a></li>
<li><span class="pkg-link" data-pkg="meteoR" data-topic="default_data_frame" data-id="view"><a href='../../meteoR/html/default_data_frame.html#method-default_data_frame-view'><code>meteoR::default_data_frame$view()</code></a></li>
<li><span class="pkg-link" data-pkg="meteoR" data-topic="default_data_frame" data-id="write_file"><a href='../../meteoR/html/default_data_frame.html#method-default_data_frame-write_file'><code>meteoR::default_data_frame$write_file()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-netatmo_outdoor-summary"></a>}}
\if{latex}{\out{\hypertarget{method-netatmo_outdoor-summary}{}}}
\subsection{Method \code{summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{netatmo_outdoor$summary(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-netatmo_outdoor-new"></a>}}
\if{latex}{\out{\hypertarget{method-netatmo_outdoor-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{netatmo_outdoor$new(dataset)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-netatmo_outdoor-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-netatmo_outdoor-finalize}{}}}
\subsection{Method \code{finalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{netatmo_outdoor$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-netatmo_outdoor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-netatmo_outdoor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{netatmo_outdoor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
