% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_df.R
\docType{data}
\name{NH0046}
\alias{NH0046}
\title{Estación NH0046 — datos meteorológicos de ejemplo}
\format{
Ver \code{\link{NH0472}}.
}
\source{
\url{https://raw.githubusercontent.com/rse-r/intro-programacion/main/datos/NH0046.csv}
}
\usage{
data(NH0046, package = "meteoR")
}
\description{
Subconjunto de observaciones horarias para la estación NH0046.
}
\keyword{datasets}
