% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_df.R
\docType{data}
\name{NH0910}
\alias{NH0910}
\title{Estación NH0910 — datos meteorológicos de ejemplo}
\format{
Ver \code{\link{NH0472}}.
}
\source{
\url{https://raw.githubusercontent.com/rse-r/intro-programacion/main/datos/NH0910.csv}
}
\usage{
data(NH0910, package = "meteoR")
}
\description{
Subconjunto de observaciones horarias para la estación NH0910.
}
\keyword{datasets}
