% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations_get_site.R
\name{observations_get_site}
\alias{observations_get_site}
\title{Retrieve hourly site observations for the past 24 hours}
\usage{
observations_get_site(site_id)
}
\arguments{
\item{site_id}{a valid met office observations site id. Use \code{observations_list_sites()} to get ID numbers.}
}
\description{
Requests observations from the met office API in xml format and returns a \code{tibble}
}
\examples{
observations_get_site(3840)
}
