% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations_list_sites.R
\name{observations_list_sites}
\alias{observations_list_sites}
\title{List available locations for past 24 hour observations}
\usage{
observations_list_sites()
}
\description{
Requests locations from the met office API in xml format and returns a \code{tibble}
}
\examples{
observations_list_sites()
}
