% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multi_Omega_adjust.R
\name{Omega_est}
\alias{Omega_est}
\title{Estimate of correlation matrix.}
\usage{
Omega_est(
  data_list = summarydata,
  ldscore.dir = ldscore.dir,
  nCores = NA,
  system_used = "linux"
)
}
\arguments{
\item{data_list}{a list of GWAS-summary-level data for exposure and k outcomes}

\item{ldscore.dir}{specify the path to the LD score files.}

\item{nCores}{The number of required cores or NA.}

\item{system_used}{The system used.}
}
\value{
A list of estimated parameters for the correlation matrix
\item{Omega}{the estimate of Omega }
\item{Omega.se}{the estimated matrix consists of the standard errors}
}
\description{
Estimate of correlation matrix.
}
