% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multi_Omega_adjust_nopar.R
\name{Omega_est_nopar}
\alias{Omega_est_nopar}
\title{Estimate of correlation matrix without parallel.}
\usage{
Omega_est_nopar(data_list = summarydata, ldscore.dir = ldscore.dir)
}
\arguments{
\item{data_list}{a list of GWAS-summary-level data for exposure and k outcomes}

\item{ldscore.dir}{specify the path to the LD score files.}
}
\value{
A list of estimated parameters for the correlation matrix
\item{Omega}{the estimate of Omega }
\item{Omega.se}{the estimated matrix consists of the standard errors}
}
\description{
Estimate of correlation matrix.
}
