% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{paginateAPI}
\alias{paginateAPI}
\title{Paginated API Call}
\usage{
paginateAPI(url, method = method, verbose = TRUE, n = 50, n_max = Inf)
}
\arguments{
\item{url}{A character string representing the URL of the API call.}

\item{method}{A character string indicating the HTTP method to be used ("get" or "post").}

\item{verbose}{Logical indicating whether to display progress messages.}

\item{n}{An integer specifying the number of results per page.}

\item{n_max}{An integer specifying the maximum number of results to fetch.}
}
\value{
The aggregated response from paginating through the API call.
}
\description{
This function paginates through the results of the provided API URL based on the specified HTTP method and returns the aggregated response.
}
\examples{
pag <- paginateAPI(url= generateURL(method = "get", ressource = "view", option = "random"),
                    method = "get")

}
