% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{post_generic}
\alias{post_generic}
\title{Generic POST Function}
\usage{
post_generic(
  method = "post",
  ressource = "external",
  option = c("cran", "doi", "instagram", "telegram", "twitter", "vk", "website"),
  package = NULL,
  doi = NULL,
  fresh = TRUE,
  handle = NULL,
  website = NULL,
  format = c("dataframe", "raw")
)
}
\arguments{
\item{method}{A character string specifying the HTTP method to be used ("post" by default).}

\item{ressource}{A character string specifying the type of resource to submit data to ("external" by default).}

\item{option}{A character string specifying the specific option for the resource ("cran", "doi", "instagram", "telegram", "twitter", "vk", "website").}

\item{package}{A character string specifying the package name (required for "cran" option).}

\item{doi}{A character string specifying the DOI (required for "doi" option).}

\item{fresh}{Logical indicating freshness (TRUE by default).}

\item{handle}{A character string specifying the handle (required for "instagram", "telegram", "twitter", "vk" options).}

\item{website}{A character string specifying the website URL (required for "website" option).}

\item{format}{A character string indicating the format of the output ("dataframe" or "raw").}
}
\value{
The response from the API call in the specified format.
}
\description{
This function performs a POST request to submit data to the specified API resource.
}
\examples{
dplyr <- post_generic(option = "cran", package = "dplyr", format = "dataframe")

}
