% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{radian}
\alias{radian}
\title{Radian}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
radian
}
\description{
Multiplicative factore used for conversion from degrees to
radians.
}
\keyword{datasets}
