% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{radiation_to_evapotranspiration_unit_conversion}
\alias{radiation_to_evapotranspiration_unit_conversion}
\title{Convert Watt/m2 to mm evaporated water per day}
\usage{
radiation_to_evapotranspiration_unit_conversion(radiation)
}
\arguments{
\item{radiation}{Radiation in W/m2}
}
\value{
radiative evapotranspiration in mm H2O per day
}
\description{
Convert Watt/m2 to mm evaporated water per day
}
