% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_example_data.R
\docType{data}
\name{raw_example_data}
\alias{raw_example_data}
\title{Unprocessed example data}
\format{
A data frame with 1802 rows and 105 variables
}
\usage{
data(raw_example_data)
}
\description{
The dataset provided is in wide format and has not been processed.
It contains missing values and duplicates, which need to be addressed prior to using MeTEor.
}
\keyword{datasets}
