% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_key.R
\name{save_api_key}
\alias{save_api_key}
\title{Register a Met Office API key to make requests}
\usage{
save_api_key(api_key)
}
\arguments{
\item{api_key}{a valid met office api key}
}
\description{
Saves a .rds format file in the current working directory to store API credentials. All API functions will fail until you have done this!
Register for a new key here: https://register.metoffice.gov.uk/WaveRegistrationClient/public/register.do?service=datapoint
}
\examples{
save_api_key(your valid key)
}
