% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{solar_constant}
\alias{solar_constant}
\title{Solar constant}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
solar_constant
}
\description{
Amount of power per area emitted by the sun which reaches the
Earth's top of the atmosphere. Given in units of W/m^2^.
}
\keyword{datasets}
