% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{solar_declination}
\alias{solar_declination}
\title{Get solar declination for given DOY}
\usage{
solar_declination(DOY)
}
\arguments{
\item{DOY}{day of year}
}
\value{
declination Solar declination in radians
}
\description{
The solar declination starts from 0 degree in the March equinox (March 20)
and increases to its maximum value of the Earth's tilt angle of 23.4
degrees in the Summer solstice (June 20/21). From there it falls again,
crossing 0 during the September equinox (September 22/23) before
approaching its minimum of -23.4 degree on the Winter solstice (December
21/22).
}
