% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{srad_to_PAR}
\alias{srad_to_PAR}
\title{Convert irradiance to photorsynthetically active radiation}
\usage{
srad_to_PAR(srad, fraction = 0.47)
}
\arguments{
\item{srad}{Average sunlight irradiance in J/s/m^2}

\item{fraction}{Portion of \emph{srad} that is photosynthetically active}
}
\value{
PAR in MJ/m^2
}
\description{
Integrates the daily average irradiation over a full day
and applies a fractional conversion factor.
}
