% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sunset_hourangle}
\alias{sunset_hourangle}
\title{Calculate sunset hourangle for given DOY at given latitude.}
\usage{
sunset_hourangle(DOY, latitude)
}
\arguments{
\item{DOY}{day of year}

\item{latitude}{Geographical latitude in degrees}
}
\value{
Hourangle in radians
}
\description{
Calculate sunset hourangle for given DOY at given latitude.
}
