% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switzerland.R
\name{swiss_coords_to_lat_lon}
\alias{swiss_coords_to_lat_lon}
\title{Convert Swiss grid coordinates}
\usage{
swiss_coords_to_lat_lon(x, y)
}
\arguments{
\item{x}{Coordinate on the Swiss grid in km.}

\item{y}{As \emph{x}.}
}
\value{
lat, lon Longitude and latitude in degrees.
}
\description{
Convert coordinates from the Swiss coordinate system LV03 into global
longitude and latidute values.
}
\note{
The Swiss coordinate system uses x to refer to the North-South axis, while
y designates the East-West axis.
}
