% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabla_resumen_temperatura.R
\name{tabla_resumen_temperatura}
\alias{tabla_resumen_temperatura}
\title{Title}
\usage{
tabla_resumen_temperatura(datos)
}
\arguments{
\item{datos}{el nombre del dataframe con las columnas id, temperatura_abrigo_150cm_maxima, temperatura_abrigo_150cm_minima}
}
\value{
dataframe con la temperatura media, desviacion estandar, maximo y minimo por id
}
\description{
Title
}
\examples{
 tabla_resumen_temperatura(NH0046)
}
