% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{terran_angular_velocity}
\alias{terran_angular_velocity}
\title{Earth angular velocity}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
terran_angular_velocity
}
\description{
Angular velocity of  Earth's orbit around the sun given in
radians/day.
}
\keyword{datasets}
