% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{terrestrial_radiation}
\alias{terrestrial_radiation}
\title{Portion of extraterrestrial radiation that reaches Earth's surface}
\usage{
terrestrial_radiation(r_extraterrestrial, relative_sunshine_duration)
}
\arguments{
\item{r_extraterrestrial}{Average extraterrestrial radiation in J/s/m^2.}

\item{relative_sunshine_duration}{Expressed as fraction of a day.}
}
\value{
Average terrestrial radiation in W/m^2
}
\description{
Implements the Angstrom-Prescott formula for estimation of terrestrial
radiation based on extraterrestrial radiation and relative sunshine
duration.
The Rietveld coefficients a and b are taken from
Rietveld, M. R. A New Method for Estimating the Regression Coefficients in
the Formula Relating Solar Radiation to Sunshine. Agricultural Meteorology
1978, 19 (2), 243–252. https://doi.org/10.1016/0002-1571(78)90014-6.
}
