% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup-datasets.R
\docType{data}
\name{visibility_lookup}
\alias{visibility_lookup}
\title{Visibility lookup}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7 rows and 2 columns.}
\usage{
data(visibility_lookup)
}
\description{
Visibility lookup
}
\examples{
data(visibility_lookup)
}
\keyword{datasets}
