% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup-datasets.R
\docType{data}
\name{weather_type_lookup}
\alias{weather_type_lookup}
\title{Weather type lookup}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 31 rows and 2 columns.}
\usage{
data(weather_type_lookup)
}
\description{
Weather type lookup
}
\examples{
data(weather_type_lookup)
}
\keyword{datasets}
