% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example data to test MeTEor}
\format{
A data frame with 180000 rows and 6 variables
}
\usage{
data(example_data)
}
\description{
An example longitudinal data set to test MeTEor
}
\keyword{datasets}
