\name{meteor-package}
\alias{meteor-package}
\alias{meteor}

\docType{package}

\title{ 
The meteor package
}

\description{
This pacakges contains of a number of meteorological data manipulation functions. 
Some of these are also available in other R packages. The context of this package is to make the functions available from dynamic simulation models of crops and crop diseases. 
}


\keyword{ package }

