% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_example_data.R
\docType{data}
\name{raw_example_data}
\alias{raw_example_data}
\title{Unprocessed sample data}
\format{
A data frame with 1802 rows and 105 variables
}
\usage{
data(raw_example_data)
}
\description{
Unprocessed sample data set, which is in wide format and contains missing values.
}
\keyword{datasets}
